Basket API configuration as a gRPC client

+++ src/Services/Basket/Basket.API/Models/ShoppingCart.cs
@@ -4,7 +4,7 @@ public class ShoppingCart
 {
     public string Username { get; set; } = default!;
     public List<ShoppingCartItem> Items { get; set; } = new();
-    public decimal TotalPrice => Items.Sum(x => x.Price * x.Quantity);
+    public double TotalPrice => Items.Sum(x => x.Price * x.Quantity);
 
     public ShoppingCart(string userName)
     {
